+++ client/src/components/auth/register-form.tsx
